fix: guard spell autolearn writes during readonly builds

+++ tswow-scripts/wotlk/std/Spell/SpellAutoLearn.ts
@@ -1,5 +1,5 @@
 import { makeMaskCell32, MaskCellWrite, MaskCon } from "../../../data/cell/cells/MaskCell";
-import { isTrinityCore } from "../../../data/Settings";
+import { BuildArgs, isTrinityCore } from "../../../data/Settings";
 import { SqlRow } from "../../../data/sql/SQLRow";
 import { spell_autolearnRow } from "../../sql/spell_autolearn";
 import { SQL } from "../../SQLFiles";
@@ -9,7 +9,7 @@ import { RaceMask } from "../Race/RaceType";
 import { Spell } from "./Spell";
 import { SpellRegistry } from "./Spells";
 
-if(isTrinityCore()) {
+if(isTrinityCore() && BuildArgs.WRITE_SERVER) {
     SQL.Databases.world_dest.read(`DELETE FROM \`spell_autolearn\`;`);
     SQL.spell_autolearn.queryAll({})
         .forEach(x=>{
